backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
google-genai
anthropic
Pillow
python-dotenv
